Decoder class for displaying names (not necessary before)
